Revert sse changes
